fixed my code

+++ src/main/resources/init_db.sql
@@ -1,6 +1,6 @@
 CREATE TABLE books (
     id INT NOT NULL AUTO_INCREMENT,
     title VARCHAR(255),
-    price INT,
+    price DECIMAL,
     PRIMARY KEY (id)
 )
